refactor(mitm-addon): split registry firewall resolution
